+++ python_modules/dagster/docs/snapshots/snap_test_doc_build.py
@@ -20237,7 +20237,8 @@ snapshots['test_build_all_docs 52'] = '''
 <dt id="dagster.execute_pipeline">
 <code class="descclassname">dagster.</code><code class="descname">execute_pipeline</code><span class="sig-paren">(</span><em>pipeline</em>, <em>environment_dict=None</em>, <em>throw_on_user_error=True</em>, <em>execution_metadata=None</em>, <em>solid_subset=None</em><span class="sig-paren">)</span><a class="headerlink" href="#dagster.execute_pipeline" title="Permalink to this definition">¶</a></dt>
 <dd><p>“Synchronous” version of <a class="reference internal" href="#dagster.execute_pipeline_iterator" title="dagster.execute_pipeline_iterator"><code class="xref py py-func docutils literal notranslate"><span class="pre">execute_pipeline_iterator()</span></code></a>.</p>
-<p>Note: throw_on_user_error is very useful in testing contexts when not testing for error conditions</p>
+<p>Note: throw_on_user_error is very useful in testing contexts when not testing for error
+conditions</p>
 <table class="docutils field-list" frame="void" rules="none">
 <col class="field-name" />
 <col class="field-body" />
@@ -20245,8 +20246,8 @@ snapshots['test_build_all_docs 52'] = '''
 <tr class="field-odd field"><th class="field-name">Parameters:</th><td class="field-body"><ul class="first simple">
 <li><strong>pipeline</strong> (<a class="reference internal" href="definitions.html#dagster.PipelineDefinition" title="dagster.PipelineDefinition"><em>PipelineDefinition</em></a>) – Pipeline to run</li>
 <li><strong>environment</strong> (<em>dict</em>) – The enviroment that parameterizes this run</li>
-<li><strong>throw_on_user_error</strong> (<em>bool</em>) – throw_on_user_error makes the function throw when an error is encoutered rather than returning
-the py:class:<cite>SolidExecutionResult</cite> in an error-state.</li>
+<li><strong>throw_on_user_error</strong> (<em>bool</em>) – throw_on_user_error makes the function throw when an error is encoutered rather than
+returning the py:class:<cite>SolidExecutionResult</cite> in an error-state.</li>
 </ul>
 </td>
 </tr>
